EliteMobs 8.2.7 - [New] Added the PLAY_ANIMATION action type for elite scripting, plays the animation for a boss model for the targets of the script - [Fix] Fixed issue which caused custom models to transition between animations very slowly - [Fix] Fixed issue where SET_MOB_AI and SET_MOB_AWARE scripts with duration would have a duration of 0 ticks - [Fix] Fixed bad target behavior on command-based elite scripts - [Fix] Fixed issue where non-default particle speed could only be integer values - [Fix] Fixed issue where elite scripts in boss files would not run if no powers were present - [Fix] Fixed bug related to zone-based scripts not working correctly - [Fix] Lowered warning rate for incorrect boss health - [Fix] Fixed error related to older mc versions trying to display the freeze effect overlay - [Fix] Fixed warning message when using locations for the target of script zones - [Fix] Fixed issue where cylinder zones would not respect the lower y boundary - [Fix] Fixed issue with tag conditions that would cause the checks to not work under specific conditions - [Fix] Fixed issue where instanced bosses did not scale health with player count correctly

+++ src/main/java/com/magmaguy/elitemobs/powers/scripts/caching/ScriptParticlesBlueprint.java
@@ -72,7 +72,7 @@ public class ScriptParticlesBlueprint {
                 case "z" -> z = parseDouble(key, value, scriptName);
                 case "amount" -> amount = parseInteger(key, value, scriptName);
                 case "particle" -> particle = parseEnum(key, value, Particle.class, scriptName);
-                case "speed" -> speed = parseInteger(key, value, scriptName);
+                case "speed" -> speed = parseDouble(key, value, scriptName);
                 case "red" -> red = parseInteger(key, value, scriptName);
                 case "green" -> green = parseInteger(key, value, scriptName);
                 case "blue" -> blue = parseInteger(key, value, scriptName);
